\name{Quantiles.TA}
\alias{Quantiles.TA}
\title{
Quantile values of the robustified statistic, TA.
}
\description{
Quantiles of the robustified statistic, TA.
They are obtained using the extensive Monte Carlo simulation with 1E08 replicates. 
}
\usage{Quantiles.TA}
\format{
  This data frame contains 97 rows and 500 columns.
}

\references{
Park, C., M. Wang, and W.-Y. Hwang (2022).
Empirical distributions of the robustified \emph{t}-test statistics.
\emph{Industrial Engineering & Management Systems}, \bold{21}(3), 432-439. \cr 
\doi{10.7232/iems.2022.21.3.432}
}

\author{Chanseok Park}

\keyword{datasets}
