%
\name{Quantiles.TB}
\alias{Quantiles.TB}
\title{
Quantile values of the robustified statistic, TB.
}
\description{
Quantiles of the robustified statistic, TB.
They are obtained using the extensive Monte Carlo simulation with 1E08 replicates. 
}
\usage{Quantiles.TB}
\format{This data frame contains 97 rows and 500 columns.}

\references{
Park, C. and M. Wang (2018).
Empirical distributions of the robustified \emph{t}-test statistics.
\emph{ArXiv e-prints}, 1807.02215. \cr
\url{https://arxiv.org/abs/1807.02215}
}

\author{Chanseok Park and Min Wang}
\keyword{datasets}

